# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/availabilities": {
      "post": {
        "operationId": "AvailabilityController_create",
        "summary": "Create a new availability",
        "description": "Creates a new availability slot for the authenticated user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Availability creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Availability created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "dayOfWeek must be a number",
                        "startTime must be a valid time"
                      ]
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Availabilities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "AvailabilityController_findAll",
        "summary": "Get user availabilities",
        "description": "Retrieves all availability slots for the authenticated user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of user availabilities retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Availabilities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/availabilities/day/{dayOfWeek}": {
      "get": {
        "operationId": "AvailabilityController_findByDayOfWeek",
        "summary": "Get availabilities by day of week",
        "description": "Retrieves availability slots for a specific day of the week (0=Sunday, 1=Monday, etc.)",
        "parameters": [
          {
            "name": "dayOfWeek",
            "required": true,
            "in": "path",
            "description": "Day of week (0-6, where 0 is Sunday)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of availabilities for the specified day retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid day of week",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Day of week must be between 0 and 6"
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Availabilities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/availabilities/{id}": {
      "get": {
        "operationId": "AvailabilityController_findOne",
        "summary": "Get availability by ID",
        "description": "Retrieves a specific availability slot by its ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Availability ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Availability retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Availability not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Availability not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Availabilities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "AvailabilityController_update",
        "summary": "Update availability",
        "description": "Updates an existing availability slot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Availability ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Availability update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Availability updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Availability not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Availability not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Availabilities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AvailabilityController_remove",
        "summary": "Delete availability",
        "description": "Permanently deletes an availability slot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Availability ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Availability deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Availability deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Availability not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Availability not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Availabilities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/availabilities/batch": {
      "post": {
        "operationId": "AvailabilityController_batchUpdate",
        "summary": "Batch update availabilities",
        "description": "Updates multiple availability slots at once",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Batch availability update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Availabilities updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "availabilities must be an array",
                        "Invalid availability data"
                      ]
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Availabilities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Manexo API",
    "description": "API documentation for Manexo application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Availability": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único de la disponibilidad",
            "example": 1
          },
          "dayOfWeek": {
            "type": "number",
            "description": "Día de la semana (0-6, donde 0 es domingo)",
            "example": 6
          },
          "startTime": {
            "type": "string",
            "description": "Hora de inicio en formato HH:mm",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "description": "Hora de fin en formato HH:mm",
            "example": "17:00"
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo de la disponibilidad",
            "example": true
          },
          "user": {
            "description": "Usuario que ofrece el servicio",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "bookings": {
            "description": "Reservas asociadas a esta disponibilidad",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Booking"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de creación de la disponibilidad",
            "example": "2024-03-20T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de última actualización de la disponibilidad",
            "example": "2024-03-20T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "dayOfWeek",
          "startTime",
          "endTime",
          "isActive",
          "user",
          "bookings",
          "createdAt",
          "updatedAt"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único del usuario",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario",
            "example": "Juan"
          },
          "lastName": {
            "type": "string",
            "description": "Apellido del usuario",
            "example": "Perez"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico del usuario",
            "example": "usuario@ejemplo.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario (encriptada)",
            "example": "********"
          },
          "avatar": {
            "type": "string",
            "description": "Avatar del usuario",
            "example": "https://example.com/avatar.jpg"
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo del usuario",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de creación del usuario",
            "example": "2024-03-20T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de última actualización del usuario",
            "example": "2024-03-20T12:00:00.000Z"
          },
          "services": {
            "description": "Lista de servicios asociados al usuario",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          },
          "availabilities": {
            "description": "Lista de disponibilidades del usuario",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Availability"
            }
          },
          "bookings": {
            "description": "Lista de reservas realizadas por el usuario",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Booking"
            }
          }
        },
        "required": [
          "id",
          "email",
          "password",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Service": {
        "type": "object",
        "properties": {
          "bookings": {
            "description": "Lista de reservas asociadas al servicio",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Booking"
            }
          }
        }
      },
...
      "CreateAvailabilityDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "number",
            "description": "Día de la semana (0-6, donde 0 es domingo)",
            "example": 6,
            "minimum": 0,
            "maximum": 6
          },
          "startTime": {
            "type": "string",
            "description": "Hora de inicio en formato HH:mm",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "description": "Hora de fin en formato HH:mm",
            "example": "17:00"
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo de la disponibilidad",
            "example": true
          }
        },
        "required": [
          "dayOfWeek",
          "startTime",
          "endTime"
        ]
      },
      "UpdateAvailabilityDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "number",
            "description": "Día de la semana (0-6, donde 0 es domingo)",
            "example": 6,
            "minimum": 0,
            "maximum": 6
          },
          "startTime": {
            "type": "string",
            "description": "Hora de inicio en formato HH:mm",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "description": "Hora de fin en formato HH:mm",
            "example": "17:00"
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo de la disponibilidad",
            "example": true
          }
        }
      },
      "BatchAvailabilityDto": {
        "type": "object",
        "properties": {
          "availabilities": {
            "description": "Array de disponibilidades a actualizar",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateAvailabilityDto"
            }
          }
        },
        "required": [
          "availabilities"
        ]
      },
...
    }
  }
}